Rewrite block refs with observableMap, which moves to compose
